Clear the clippy findings the linkage work brought with it
Two useless `i64::from` on values already `i64`, and a reasoned allow on
`structured_features`, which is one block per option-shaped fact and reads worse
split. The three warnings left in the workspace all predate this branch.

+++ crates/ti4-engine/src/objectives.rs
@@ -411,8 +411,8 @@ impl<'a> Position<'a> {
                     .is_some_and(|name| !real.contains(name))
             })
             .map(|planet| match kind {
-                crate::production::Spend::Resources => i64::from(planet.resources()),
-                crate::production::Spend::Influence => i64::from(planet.influence()),
+                crate::production::Spend::Resources => planet.resources(),
+                crate::production::Spend::Influence => planet.influence(),
             })
             .sum()
     }
